Backend/requirements.txt
fastapi
uvicorn
google-auth
pymongo
dnspython
motor
pydantic
aiokafka
requests
python-dotenv
python-jose[cryptography]